main.py
import datetime
import os.path
import subprocess
import pickle
import tkinter as tk

import cv2
import face_recognition
# to converty the img into pillow
from PIL import Image, ImageTk

# for import test function in file test
import util


class App:
    def __init__(self):
        self.main_window = tk.Tk()
        self.main_window.geometry("1200x520+350+100")

        self.login_button_main_window = util.get_button(self.main_window, 'login', 'green', self.login)
        self.login_button_main_window.place(x=750, y=200)

        self.logout_button_main_window = util.get_button(self.main_window, 'logout', 'red', self.logout)
        self.logout_button_main_window.place(x=750, y=300)

        self.register_new_user_button_main_window = util.get_button(self.main_window, 'register new user', 'gray',
                                                                    self.register_new_user, fg='black')
        self.register_new_user_button_main_window.place(x=750, y=400)

        self.webcam_label = util.get_img_label(self.main_window)
        self.webcam_label.place(x=10, y=0, width=700, height=500)

        self.add_webcam(self.webcam_label)

        self.db_dir = './db'
        if not os.path.exists(self.db_dir):
            os.mkdir(self.db_dir)

        self.log_path = './.log.txt'

    def login(self):
        unknown_img_path = './.img.jpg'
        cv2.imwrite(unknown_img_path, self.most_recent_capture_arr)

        output = str(subprocess.check_output(['face_recognition', self.db_dir, unknown_img_path]))
        name = output.split(',')[1][:-5]

        if name in ['unknown_person', 'known_people']:
            util.msg_box('INVALID', 'Unknown')
        else:
            util.msg_box('Welcome!', 'Welcome {}.'.format(name))
            with open(self.log_path, 'a') as f:
                f.write('{},{}\n'.format(name, datetime.datetime.now()))
                f.close()
        os.remove(unknown_img_path)

    def logout(self):
        self.main_window.destroy()

    def register_new_user(self):
        # create another window for register
        self.register_new_user_window = tk.Toplevel(self.main_window)
        self.register_new_user_window.geometry("1200x520+370+120")
        self.accept_button_register_new_user_window = util.get_button(self.register_new_user_window, 'ACCEPT', 'blue',
                                                                      self.accept_button_register_new_user)
        self.accept_button_register_new_user_window.place(x=750, y=300)

        self.try_again_button_register_new_user_window = util.get_button(self.register_new_user_window, 'Try Again',
                                                                         'red',
                                                                         self.try_again_register_new_user)
        self.try_again_button_register_new_user_window.place(x=750, y=400)

        self.capture_label = util.get_img_label(self.register_new_user_window)
        self.capture_label.place(x=10, y=0, width=750, height=500)

        self.add_img_to_label(self.capture_label)

        self.entry_text_register_new_user = util.get_entry_text(self.register_new_user_window)
        self.entry_text_register_new_user.place(x=750, y=150)

        self.text_label_register_new_user = util.get_text_label(self.register_new_user_window, 'INPUT USERNAME :')
        self.text_label_register_new_user.place(x=750, y=70)

    def start(self):
        self.main_window.mainloop()

    def accept_button_register_new_user(self):
        name = self.entry_text_register_new_user.get(1.0, "end-1c")

        cv2.imwrite(os.path.join(self.db_dir, '{}.jpg'.format(name)), self.register_new_user_capture)

        util.msg_box('Success!', 'User was registered successfully !')

        self.register_new_user_window.destroy()

    # to exit the window and go to the main window
    def try_again_register_new_user(self):
        self.register_new_user_window.destroy()

    def add_webcam(self, label):
        if 'cap' not in self.__dict__:
            self.cap = cv2.VideoCapture(0)

        self._label = label
        # to add the webcam in the label
        self.process_webcam()

    def process_webcam(self):
        ret, frame = self.cap.read()
        # put the frame in this variable then we gonna converted
        self.most_recent_capture_arr = frame

        # for convert color to rgb
        img_ = cv2.cvtColor(self.most_recent_capture_arr, cv2.COLOR_RGB2BGR)

        # to convert into pillow
        self.most_recent_capture_pil = Image.fromarray(img_)

        imgtk = ImageTk.PhotoImage(image=self.most_recent_capture_pil)
        self._label.imgtk = imgtk
        self._label.configure(image=imgtk)

        # repeat the process every 20s
        self._label.after(20, self.process_webcam)

    # for adding image to the label (just image)
    def add_img_to_label(self, label):
        imgtk = ImageTk.PhotoImage(image=self.most_recent_capture_pil)
        label.imgtk = imgtk
        label.configure(image=imgtk)

        self.register_new_user_capture = self.most_recent_capture_arr.copy()


if __name__ == "__main__":
    app = App()
    app.start()
